Adds a script to fix all SVG files in files/Tex.
Some generated SVG files might use quadratic instead of cubic curves.
This causes and error with manim as it falsely interprets them and, as a
result of that, causes them to render slightly wrong.

This script, fix_svgs.sh, manually fixes these using cairosvg. This is a
haskish solution, but it does work. You'll need to run the script and
re-extract your scenes.A